Deleting my attempt at making a generic Anki ConfigTree
Why you ask? It's too much of a PITA. The types are self referential,
so you really want an operator[] that returns a ConfigTree& reference
so you aren't copying. but the underlying structure has type
Json::Value, so you can't return a ConfigTree reference on a
Json::Value object. The choices are to copy it every time, or to go
and convert the whole tree into ConfigTree once and return refs to
that, or maybe to convert everything to void*'s. Not pretty