Created robot_advertisement_controller to emulate BTLE advertising when using simulated robots. Any world using CozmoBots should include a RobotAdvertisementController.
Robots register to the service by connecting to UDP server on port ROBOT_ADVERTISEMENT_REGISTRATION_PORT and sending a RobotAdvertisementRegistration message.

A basestation that wants to view available robots connects to the UDP server on port ROBOT_ADVERTISING_PORT and sends any amount of data to register as a client to receive advertisements.  Advertisements come in the form of a RobotAdvertisement message.