Fixed issue with absLocalizationUpdates that was causing robot to report incorrect poses back to basestation. (This in turn was causing oscillations in observed mat marker pose even when the robot was motionless.)
Example of what was happening:
Assume robot is not moving, but looking at a mat marker.
At t=15, a keypose is received for t=10 in an absLocalizationUpdate msg.
The corrected pose is computed for t=15 by taking the diff of historical pose at t=10 and the current pose at t=15 and applying it to the key pose.
The historical pose at t=15 is not updated, but subsequent poses are identical to the newly computed pose.
At t=20, a key pose is received for t=15. A corrected pose is computed for t=20 in the same manner as before. We expect the diff transform to be identity, but it's not because the historical pose at t=15 is different from the current pose at t=20.

What we need to do is ignore all absLocalizationUpdates unto AND INCLUDING the last absLocalization update time.