Rework code structure and linker scripts significantly to get everything to build and run together.
Some whole libraries are now put entirely in flash rather than putting the ICACHE_FLASH_ATTR
decorator on each function. For C++ this is especially nessisary since it does not fit
otherwise.

The Espressif does not support unalighned reads and writes so everything needs to be done
aligned. I had to reword embedded reliable transport to fix this.