Pass Unity executable path to unity.py script
We need to pass the unity executable path specified as an optional configure option
to unity.py in order to use the desired Unity version for builds.  Without this, it
is not possible to use a version of Unity that is not installed in
`/Applications/Unity/Unity.app`.  We need multiple version support for sharing build
agents and allowing engineers to work on both Drive & Cozmo.