Fixed behavior manager bug causing animation skips
Before, when switching behavior choosers, the old behavior could still run for up to a second. In the case of
the bored animation, the old faces behavior could queue an action which would cancel the animation.

This fix forces the new behavior chooser to select a behavior right after it gets set