Mood and Progression messages now allowed in union only
Update to cozmo-engine enforces that messages that are intended to be packed in a union can no longer be sent outside of that union. Caught one usage outside of that (engine would have silently ignored this message) so fixed usage in Robot.cs. Engine changes (mostly unrelated but i need the latest one):

* lib/anki/cozmo-engine da59b38...c932a7c (5):
  > Merge pull request #515 from anki/BehaviorGroupEnablingViaCLAD
  > Merge pull request #513 from anki/merge-demo-into-master
  > Merge pull request #516 from anki/andrew/faceRec-WIP
  > Merge pull request #514 from anki/python3_depends
  > Make the k02 use much less stack when parsing clad messages. Improves stability but not fixed completely.