Behaviors can now be enabled/disabled at any granularity
Behavior group enabling is now supported on a per-behavior basis - so you can e.g. disable a group, but then individually enable some of the behaviors in that group, or disable a group but then enable a sub-group from that.
To support this, each behavior tracks whether it is choosable or not.