BodyMotion StopMessage Bug Fix
Don't send a "Stop Body Motion" animation message from BodyMotionKeyFrames if the keyframe's duration takes it to within a single animation sample length of the next keyframe in the track. This is accomplished via a specialization of the AddKeyFrameToBack method for the BodyMotion animation track, which checks the previous keyframe to see if it ends too close to the keyframe being added and disables the stop message if so. 

COZMO-1208
COZMO-1499