COZMO-1782 Put DASConfig.json in common location
Currently DASConfig.json lives in /unity/ios, where the Android
build has no access to it. This change moves it into the
StreamingAssets folder, where Unity will package it in a location
that we can access it on either platform. We also have to pass the
root of the StreamingAssets folder to C++ (currently we tack
"cozmo_resources" on the end first), so that C++ knows where to
find DASConfig.json.