Fix uses of auto in C++ code making spurious copies
Many places in our C++ code are using the 'auto' keyword in
situations where it would result in unnecessarily copying objects
that can instead be accessed via a reference. This commit changes
as many instances of this as I could find (doing a regex search to
find instances of 'auto' not followed by an ampersand, and checking
to see if the type being assigned is something simple like a
primitive type or a pointer versus a more complex type we'd prefer
to avoid copying) to use a `const auto&` reference instead.