Three robot fixes in one!
1) COZMO-2017: Stopped moving message was not being reported for blocks when they were at rest on one axis.
The amount of variance in accelerometer values seems to have gone up since EP2 (10% error in spec sheet) so I changed the way motion is being detected. It's not any hacker than what was already there, but we should probably do this differently once accelerometer values are being fed up to engine.

2) COZMO-2026: Robot pitch angle not updating when head is moving.
Made it so that now it is updated using gyro integration, but beware since the pitch reading is pretty noisy when the head is moving. (Whenever the robot stops moving for a brief period the pitch is corrected to an absolute measurement determined by accelerometer.)

3) COZMO-2025: Robot can keep driving if animation hangs
Now stopping wheels whenever the animation buffer is starved.