Fixture c19 - 2482->2481 and use ESP "EMC" firmware.
At the test range, we found out we needed our high channel to be 2481,
not 2482.

We also learned the ESP "EMC" firmware build has lower harmonics than
the "RF" firmware build.  I have no clue why, but it's probably output
power related.

We found a hint that EMC firmware is meant to work with the kind of
matching network we have.