Admire stack fixes (#1221)
* COZMO-2145 Crash in pickupObjectAction

Cancel was not correctly prepping actions that haven't started running for completion.

* COZMO-2138 Cozmo plays the angry animation multiple times
COZMO-2139 App does not transition to scene 6 after knocking over stack

If any of the actions fail when in the knocking over stack state, the behavior would stop running and then immediatly start running again ultimatly causing Cozmo to play the angry animation again. Fixing this by transition to another state on action failure that just tries to knock of the stack.

It is possible that Cozmo thinks the objects in the stack are not actually where they are, this means that when flipBlockAction tries to clear the stack objects are not considered on top of each other. Fixing this by having the behavior clear all lightcubes when knock over is successful