COZMO-2167: Investigate a ping solution for the connect button (#125)
* Reimplemented the ping functionality to use coroutines like it is suggested in the documentation.
After a successful ping, now we continue pinging, although less often, to catch the case when we lose connectivity to the robot.

* Commented out the Ping functionality because it seems to be leaking threads. The issue has been reported to Unity. In the meantime, we'll use the WiFi IP address to know if we are in the robot subnetwork.
This allowed me to find a bug on the function that finds the WiFi address. On devices with multiple interfaces (like a laptop), it was returning the first IPv4 address which may not be the right one. Similar code on engine look for the en0 interface so I'm following suit here.