COZMO-2530 Fix webotsTest.py being killed if ran in the same directory (#301)
Previously the kill process function had to be modified such that it doesn't kill another script (AnimationTests.py) importing and reusing the kill process function that's in webotsTest.py. It was modified to kill any process that matches the regex `[w]ebots` but ignore anything with the project/build-scripts/webots/ path in the name. If the script is called when you're already cd'd into webots/ instead of calling project/build-scripts/webots/webotsTest.py the ignore doesn't work. The command is modified to ignore anything with `.py` now.