Added tools/debug/autoDebugEngine.sh
This script will run in the background and continuously connect lldb as
soon as it notices that the game engine is running. If the engine exits
cleanly (e.g. you quit or reload webots), it will wait for it to start
up again. If there is a crash, it will break in lldb (command line only)