Allow more specific method parameters for C# CLAD subscription (#341)
After @pohung's latest change, subscribing to receive CLAD
messages in C# required passing in delegates taking a single
`object` parameter, which would then have to be casted to the
appropriate message type in order to read data from the message.
I talked to him about using templates to allow for the
subscription of methods that accept the full message type as their
parameter (removing the need to perform a cast from an `object`)
and came up with this change to do so.

Since the callback-managing code is a little bit hefty, I
outsourced it to a new class (`CallbackManager`) with a simple
interface which `RobotEngineManager` now owns an instance of.

One (potential, depending on your perspective) downside is that it
seems that when you create a delegate with an implicit `this`
target to act on by passing a method into a context where you
expect it to be assigned to an `Action<T>`, like this:
```c#
  public void MyFunc(SomeType asdf) {}
  public void SubscribeToThings() {
    RobotEngineManager.Instance.AddCallback(MyFunc);
  }
```
C# refuses to infer the type of `T`, no matter how obvious you
make it. This is probably the result of anonymous functions (which
I would guess the above code generates?) not having a clear,
defined type until you explicitly assign them to a delegate with a
known type of something. So, what that means is that calls to
`AddCallback` have to explicitly specify the type being subscribed
to:
```c#
  public void SubscribeToThings() {
    // have to use this syntax instead
    RobotEngineManager.Instance.AddCallback<SomeType>(MyFunc);
  }
```
On the bright side, this does make it explicitly clear what type
of message is being subscribed to at the point of subscription,
and allows for some level of type safety (you can't pass a method
that takes a parameter of a type unrelated to `T`).

This also removes the CLAD emitter modification that added
accessors for the `System.Type` of a message, since it's no longer
used.