COZMO-709 Add image latency tracking (#391)
* COZMO-709 Add image latency tracking

One of the important components for detecting when our connection
to a robot has soured is knowing how long it is between receiving
images from it, as well as if it's been awhile since we received
one. This adds the tracking of how often we're getting images, as
well as how long it's been since we last received one, and sends
it to the UI in our latency message. This also adds average image
frequency and the most recent image frequency to the debug latency
display that will pop up if lag to the robot gets bad.

General tendencies I've observed so far:
- Since image frequency is measuring time between engine updates
  in which we received an image, it tends to be a multiple of 60ms
- Time between receiving images is usually either 60ms or 120ms,
  with the average hanging out around 70-90ms
- It's not terribly uncommon to hit the occasional spurt where we
  go about 500ms without getting an image from the robot

Going forward, QA should be able to use this to identify if poor
performance from the robot correlates with either the wifi latency
or with these drops in image frequency, and if so, what numbers
define an acceptable range of performance.

* PR comments