COZMO-2708 Refactor DAS initialization code (#445)
The difference between the support offered by each platform's
version of the DAS library is staggering, which proved to be a
supreme pain in the ass when it came to setting things up for
Cozmo. The iOS das lib had Objective-C implementations for
everything you'd ever need to get from your iOS device and would
set common globals and print common events on bootup. In contrast,
the Java das lib does basically nothing except upload logs to the
server - everything else we would like to have was implemented
separately in OverDrive, in app-specific code.

This change aims to ease the pain of using DAS across multiple
platforms by bringing all this platform-specific info we need
(like, device ID, OS version, etc) under a single interface,
`IDASPlatform`. Implementations are provided for iOS (which moves
about all of its code in from the former `DASClientInfo.m`) and
for Android (which copies most of the functionality over from
`OverDriveActivity.java` in the OD project). When the C++ engine
is booted up, it creates an implementation instance and passes it
to DAS to store as its global platform, and to use to set the
aforementioned common globals/events in delightfully platform-
independent code.