COZMO-2643 face tracking test (#464), COZMO-2775 Standardize UiGameController Units
The test moves a face laterally across the camera view in the y direction and then again but in all three dimensions at once and a bit faster.

Also refactored how UiGameController accesses the light cubes. GetLightCubeById is implemented so that every function that needs to get a reference of a lightcube by id doesn't have to implement their own loop through the vector of light cubes.

CycleVizOrigin is also cleaned up to reduce code duplication and be generally more readable.

Pose translation units for all objects are standardized to millimeters in UiGameController.