COZMO-2166. BlockPool improvements (#611).
Greyed out buttons are now still interactable so cubes can be removed from the block pool even when they are not available.
The buttons are now moved between the available and enabled lists on button click to better reflect their status in the pool.
Added an Update function to solve the case when the block pool is opened before the Robot class has all the information about the cubes. In this case, the buttons would remain with the "Unknown" value for the type.