COZMO-2220: Detect Markers in Low Light with Cube LEDs on (#620)
This PR addresses problems detecting markers in low light when Cube LEDs were on. The LEDs being on affects the local contrast of the image, causing binarization to fail at the very beginning of the detection algorithm. Specifically, the fiducial's connected component would bleed out into the white area of the cube because, relative to the bright LED, the white was binarized to "black".

The solution is to use pretty heavy pre-processing via Contrast Limited Adaptive Histogram Equalization (CLAHE). Many parameter settings and approaches were tried (including processing every image twice: with and without CLAHE applied) -- thus all the commits -- but eventually I settled on a single round of processing with a fairly aggressive preprocessing. In the process I added / changed several console vars.

To test and prevent regression, this PR also includes a new engine unit test with 50 "hard" images. With this PR, we achieve detection in 49 of the 50 images. (I could only get the final one by using the double processing approach, but that seemed like too big of a performance tradeoff.)

I also made some changes to (hopefully) help with performance, but had a very hard time measuring gains.

More details:
* Fewer pyramid levels, but starting at larger kernel (should be faster). Plus specialized 3-level function for detector for speed. (One pyramid level creates three layers.)
* Fix bug in manual kmeans initialization and reduce to a single try instead of two (should be faster)
* Rewrote illumination normalization filtering step in the detector to use a (separable) box filter instead of a single non-separable specialized center-surround filter. (should be faster, based on profiling)

Unrelated stuff:
* (COZMO-2450 / COZMO-3251) Make quaternion assertion into a warning until we can figure out the root cause
* Add Save method to our Image class
* Add drop shadow option for drawing text on images
* Some fixes for printing unsigned timestamps
* Add DroppedImageStats class for tracking dropped images
* Reworked some of Vision::Profiler class (add scoped TicToc object -- which I used in working on this, but that usage is not in the final commited version, alias Resolution and ClockType)
* Adding Get_<NamedPoint> helper methods to Quad class