COZMO-2785 Android background service for uploading (#651)
This provides the Android implementation for giving the app
notifications from the OS when internet is available, so that the
app can upload DAS logs and whatever else we need.

Unlike iOS, where the OS already has a background fetch service
that performs a similar function to what we need, we have to
implement our own background service on Android to run every now
and then and see if internet is available. Since this background
service continues to run even when the app is closed, it does not
call directly into Cozmo (which would cause a C++ link error when
the app is not running), but rather sends a broadcast back to the
app to make the native call into the engine.

This also adds a super basic Java build system into the
`project/android` folder for building Cozmo-specific Java code. On
a related note, I really hate gradle.