Fix stack overflow crash in Android HTTP adapter (#868)
When sending a large (2mb) file through the HTTP adapter, this
C-style variable size array winds up blowing through the stack.
OD (the source of this code) must not be sending very large data
with it. I believe it's not even necessary to have this temporary
array, since what documentation I could find on
`SetByteArrayRegion` seemed to imply that the data passed in is
copied, so there's no reason we can't just do that straight from
the source buffer.