GameMessagePort.PushToGameMessage.send_buffer_full (#860)
GameMessagePort.PushToGameMessage.send_buffer_full happen half a million times per week. but we have very little exposure to when and why it happens. I have added asserts that should kill the app if this error occurs in our TC built apps. Hoping to catch this error as it happening.