Fixed bug in BlockWorld::FindObjectOnTopOf
Problem was that we were applying ABS to vectors incorrectly. We rotated
tolerance first, then took the ABS. If, e.g. the y coordinate after the
rotation was nearly zero, then the check was constantly failing.

The solution is to wait to apply the ABS until everything is in the same
frame. We rotate the distance between the points back into the
axis-aligned frame, then do the check there