Add smart handle wrapping local Java object references (#952)
When writing my most recent PR, I hit a case where an early return
out of a function would have leaked a local reference if I didn't
duplicate code to free that reference that already existed at the
end of the function. Since duplicating code sucks and we have the
magic of C++ at our disposal, this change adds a wrapper object
that behaves somewhat like a `unique_ptr` (and uses one
internally) and frees the local reference it's bound to in its
destructor, using the `JNIEnv` it's passed in construction.