COZMO-3881 Offload remote DAS appender's stringifying to thread
Like the previous change to the iOS local appender, this change
takes some string creation that was done on the main engine thread
by the DAS remote appender, every time a log line is called, and
moves it to instead run on the DAS queue, copying the data
necessary to create it into a lambda. According to the profiling
I've been doing, this change combined with the previous change
reduces the average time to log something in DAS  (in release
builds) from ~3.5ms to ~1.5ms.