COZMO-3881 On device, use direct channel for viz comms to Unity
As mentioned in the first change in this series, sending messages
on sockets sucks and is super painful. Removing the use of
`VizManager`'s `_vizClient` on device builds addresses half of its
performance issues, and this addresses the other half.

Similar to the change made awhile ago to change C++ <-> C#
communication to use direct method calls instead of sockets, this
adds a new `GameMessagePort` instance to the Cozmo engine instance
for `VizManager` to send messages with, only on device builds.
This also adds a corresponding implementation of the direct
communication channel (derived from `RobotDirectChannel`) for Viz,
with some modification. There's a fair amount of code re-use that
feels a little gross but since the Viz system is using this a bit
differently (messages only go one way, and the Viz UDP channel
contains a method to blow away queued messages that isn't part of
the shared `ChannelBase` hierarchy) and since this isn't a shipping
feature, it probably isn't worth making this play nice with other
direct channel implementations.