COZMO-3881 Put iOS DASLocalAppender string formation on a queue
DAS is sometimes called several times per frame, and if our
logging isn't as fast as possible then it can really slow us down.
Concatenating a string together from several pieces of data is a
particularly slow process, and copying the data needed to form it
into a block (to be executed on a queue) saves the main thread
about half the time that it previously spent doing the
concatenation itself.