COZMO-4277: FaceEnrollment Premature Timeout
Avoid situation where face enrollment is successful, but times out during the final animation, after storing the name to NVStorage. So Cozmo actually learned your face, but Unity thinks the process failed and doesn't put your name in the list until app is restarted.  Hack solution until we move all the animations to Unity: just double the timeout once we get to the Finishing state and have an animation to play. (#1138)