COZMO-4160 Lights on cubes don't turn green (#1136)
* COZMO-4160 Lights on cubes don't turn green

Fixes an issue where if game sent a SetLights command while LightsComponent was enabled it would override the current light state. Makes it so that if LightsComponent is enabled game is unable to set the cube lights.

Fixes issue where game was not reenabling the LightsComponent on game quit.

Fixes an issue with restoring the previous light state when the light state should change while LightsComponent is disabled.