COZMO-2142 Cozmo is stuck while claiming cube (#1301)
* COZMO-2142 Cozmo is stuck while claiming cube

The issue was we were considering a far away predock pose even though we were basically ignoring it. This meant that we were using the wrong marker to track/dock with (instead of the marker we are currently looking at) so the dock action would fail due to "too long without docking error signal" since we were never tracking the right marker.