Real fix for COZMO-1500: Flipping on back during quicktap (#1381)
When animation stops it sets the speed of the lift to 0, but in an effort to be clever past-Kevin made it so that it would ramp down to 0 based on the current speed, max speed, and acceleration. Because the lift moves so fast during the downstroke of the tap, the projected position for a smooth stop is at/below the lowest lift position. Now I'm just instantaneously setting the target position of the controller to the current height of the lift.