COZMO-4511 Allow video skipping in dev builds (#1383)
To allow for faster use of the app in testing, in dev builds only
(currently defined by `#if ENABLE_DEBUG_PANEL`), tapping the
screen three times while a video is playing will skip it.

Initially I attempted to call `Stop()` on the media player to
accomplish this, but that effectively pauses the video and did not
trigger end-of-video behavior. Seeking to the end, as in this PR,
did the trick.