COZMO-4637 COZMO-4629 Fix crash getting possible poses (#1405)
Normally, when `DriveToObjectAction::GetPossiblePoses` is called,
it returns failure for cases where the pose vector passed in can't
be filled. Following code in `DriveToObjectAction::InitHelper`
checks that return value to make sure it succeeded before trying
to access the first value in the pose vector.

However, when the action is set to have its possible-pose-getting
function wrapped by `IBehavior::UseSecondClosestPreActionPose`,
the return value of `DTOA::GetPossiblePoses` is discarded and
success is always returned, even if a failure happened and the
returned pose vector was empty. That situation led to a crash
when `InitHelper` then saw a successful result and tried to use
the empty vector.