Refactor SDK bootstrap routines
connect() now accepts either a synchronous or asynchronous function;
connect_async() has been removed.

connect_on_loop allows the more advanced user to establish a connection
on an existing loop and integrate it with a larger program.

Removed a lot of options for connect that were unlikely to be useful in
favour of simplicity for the common case of just wanting to start a
program.

Cleaned up connection lost and program termination handling so that
there should no longer be warnings about unfinished tasks on a clean
program exit (still a bit more work to do there).

Losing the TCP connection to the device will trigger connection shutdown
and raise an exception in any calls that are blocked waiting on an event
handler to complete, preventing loops or hung programs.