COZMO-3868 Keep Unity running after iOS inactive notification (#1435)
So, this is a weird/gross change and I'm not sure how I feel about
it. The root issue here is that when Unity gets a
`applicationWillResignActive` notification from iOS, it calls our
PauseHandler in C# and shuts everything down (pauses itself
indefinitely, until receiving `applicationDidBecomeActive`). The
JIRA ticket referenced here states that we'd like to avoid exiting
mini-games if possible when bringing up the control center, but
the code that drives those mini-games in C# cannot run while Unity
is paused.

As such, the only way I can see to keep the game running is to
keep Unity from pausing itself. This means using our subclassed
app controller to suppress the `applicationWillResignActive`
notification when receiving it, and instead send that notification
to Unity when we receive the `applicationDidEnterBackground`
notification which tells us the app is backgrounded. If the user
brings up the control center and closes it, Unity never receives
anything at all, and it will only pause itself when the app is
actually backgrounded.