COZMO-3400 Increase limit of DAS log files on disk (#1528)
This ticket describes crash logs being uploaded that have no
associated DAS logs for the reported apprun. I looked closely into
what HockeyApp is doing on iOS and it seems like its behavior here
is correct:
- app crashes
- at start of next apprun, HockeyApp generates a crash report,
  which we populate with the previous apprun
- if the report can't be uploaded, it's stored on disk until all
  pending reports can be uploaded together

Therefore, my theory about the cause of this issue is that our DAS
logs on disk are essentially a 4MB ring buffer, which can
eventually fill up and start overwriting previous logs. This was
never a problem on OD because we didn't commandeer the user's
internet connection, but long periods of being connected to Cozmo
on wifi means that lengthy Cozmo use will cause older logs to be
overwritten before they can ever be uploaded.

This change increases the size of each DAS log file, as well as
the total number of files we store on disk, by 2.5x and 4x
respectively, leading to a 10x increase in the effective size of
the DAS ring buffer on disk. This should significantly decrease
the issue of older apprun logs being deleted before upload.