COZMO-4068 Override alert sound for sleep modal (#1551)
* COZMO-4068 Override alert sound for sleep modal

AlertView had a hard-coded open sound that
played on Awake. Because the sleep modal needed
to have a different sound on open, this change
adds an open sound to BaseView (default invalid),
adds the hard-coded sound to AlertView, and overrides
the sound for the sleep modal specifically.

* fix GameObjectType in AudioEventParameter

AudioEventParameter was not setting the
GameObjectType when serialized on a GameObject.
It turned out that this value does not need to
be serialized anyway, and can be derived from
the EventGroupType, so it has been removed and
replaced with a getter method.

AudioEventParameter.VOEvent was removed because
it is unused.

Currently, the only valid EventGroupTypes for
AudioEventParameter are UI and SFX, so everything
else will return invalid when retrieving the GameObjectType.