COZMO-3393: Seeing object on lift (#1696)
Fixes an issue with seeing an object we think is on our lift at similar poses which causes a new object to be added to blockworld meaning there will be multiple objects of the same type in the world. This is at least one case that could lead to the Robot.HandleActiveObjectConnectionState.MismatchedIDs error in this ticket.