COZMO-5126 allow overrides for unlocks in freeplay (#1827)
Late breaking feedback came in that Cozmo isn't smart / deliberate
enough in freeplay during the first app session (before anything is
unlocked). The solution is to allow Cozmo to use stacking and rolling in
freeplay, even though they are still technically locked (so we don't
mess with the progression tuning / flow). This means that the core
upgrades still need to appear locked in the UI, but the robot should be
able to perform the actions (but the user can not spark them).

This is implemented by adding freeplay overrides to unlocks. Now, when
asking if an unlock is unlocked, you also specify whether or not you are
asking "for freeplay purposes" or not. I defaulted it to "not freeplay"
even though it turns out this is never used (I wanted to be really
explicit about where this change would take effect).

If the progression component is asked about an unlock that is in the
data-driven "freeplayOverrides" list, it will lie and claim that it is
unlocked for freeplay purposes, even if it isn't really unlocked.

Longer term, if we want to keep functionality like this, we should
probably split the idea of "can Cozmo do X" from "is X unlocked in the UI"