COZMO-5204 Fix scoring for PlayAlone and cooldowns for AI goals
WantsToPlay was not being set if no games were available, so the robot
never wanted to play alone before games were unlocked. Solution is to
not score PlayAlone based on WantToPlay.

Added code to goal evaluator to use a much shorter cooldown if the goal
only got to run for a very short period of time. This catches the case
where a goal gets selected, but has no behaviors to run.

Don't allow the PlayAnim behavior to be resumed. This way we can move
to another goal much more quickly after an interruption such as
acknowledge object. Also, doesn't generally make sense to repeat an
animation from the beginning after an interruption