COZMO-5319 Make SDK connect loop use consistent
Synchronous functions passed to connect will cause a new thread to be
created, and a new loop to be created for that thread.  The thread calls
set_event_loop so that calls to asyncio.sleep, etc default to using the
right loop.

This patch enables similar behavior for asynchronous functions; if the
thread calling connect already has a default loop then that's used, else
a new one is created and made the default for the thread.

If the user needs more explicit control over which loop is used, they
should use connect_on_loop() instead.

Additionally, expose which loop each object (CozmoConnection, Robot,
etc) is using via a read-only loop property on the base class.