COZMO-5325 spark refund on failure
Sparks will now refund and price will not increase if the associated behavior objective is not completed successfully. Reward will still be earned if the spark completes (success or failure). If the player cancels the spark, price will not increase, they will not get a refund, and they will not get a reward