COZMO-4430 Make Android file uploads memory efficient (#2114)
Currently, when uploading files using the `HttpAdapter` for
Android, we read the entire file into a buffer in C++ and pass that
buffer to Java (which copies the memory). For large files, this can
wind up allocating a large amount of memory.

This change instead detects a file upload as an upload request
with an empty body and a filename passed in, and accordingly does
a couple things differently:
- creates an input stream from that file instead of from the
  request body string
- streams data from the file into the output stream 64k at a time

Testing this with very large files (192m) on a crappy Android
device, I got a C++ crash with the current master implementation
but had it work (with the caveat that I had to increase the
socket timeout console var to make that happen) with these changes.