COZMO-5492 Cube lights during quick tap are stuck (#2100)
* COZMO-5492 Cube lights during quick tap are stuck

The issue was caused by disconnecting and reconnecting to the cube. SpeedtapGame keeps a reference to the PlayerBlock and CozmoBlock, when either of the blocks get removed from LightCubes dictionary then they are kept around as garbage references so any calls to their SetLEDs functions will not work. To fix this I have replaced all the block.SetLEDs calls with LightCubes[block.ID].SetLEDs. This works because eventhough the block references are garbage the IDs are still correct.

The correct way to fix this would be to handle block disconnection and reconnection at the speedtap game level and invalidate PlayerBlock/CozmoBlock

* Making SetLEDs during speedTap and accessing Cozmo/Player block safer with existence checks